requirements.txt
# AI Computer Control Agent - Dependencies

# OpenAI API for vision and planning
openai>=1.12.0

# Environment variables
python-dotenv>=1.0.0

# Screenshot capture
mss>=9.0.0
Pillow>=10.0.0

# OCR
pytesseract>=0.3.10
paddlepaddle>=2.5.0
paddleocr>=3.3.0
paddlex[ocr]>=3.3.0
opencv-contrib-python>=4.10.0.84

# Mouse and keyboard control
pyautogui>=0.9.54
pynput>=1.7.6

# Windows-specific (for better input control)
pywin32>=306; sys_platform == "win32"

# Data validation
pydantic>=2.0.0

# Utilities
numpy>=1.24.0

# For colored terminal output
colorama>=0.4.6
rich>=13.0.0
